move different navigators into their own files
